+++ packages/ui/src/configurations/index.ts
@@ -3,3 +3,4 @@ export * as IMAGES from './images';
 export * from './variables';
 export * from './http';
 export * from './resources';
+export * from './kernel';