+++ shared/core/src/lib/utils/es-toolkit.ts
@@ -1,10 +1,13 @@
 export {
   assignIn,
+  dropRight,
   endsWith,
   isEmpty,
   isFunction,
+  join,
   merge,
   partition,
   pickBy,
+  split,
   toUpper,
 } from 'es-toolkit/compat';