+++ shared/framework-kernel/src/lib/main.ts
@@ -1,0 +1,9 @@
+import type { KernelOptions } from '@/declarations';
+
+import { RouterUtilities } from './utilities';
+
+const initializer = (options: KernelOptions) => {
+  RouterUtilities.initialize(options.router);
+};
+
+export { initializer };